requirements.txt
torch
torchvision
matplotlib
numpy
Pillow
tqdm
wandb
opencv-python
scikit-learn